requirements.txt
tensorflow
imageio
matplotlib
black
isort
pandas<2.0
parameter-sherpa
corner
tf2jax
numpyro
